Ut 1606 rebase on tk config default2 v1.2.1 (#26)
* SG-9477 adds launchfolder app to PublishedFile environment (#46)

Adds the tk-shotgun-launchfolder app to the PublishedFile environment, due to the app's recent support for the entity type. This involved separating the version and publishedfile environments, as the app does not support versions.
closes #46

* SG-9252: Adds tk-desktop2 to the same environments that tk-desktop is in. (#48)

Adds tk-desktop2 to the same environments that tk-desktop is in. This is in support of Shotgun Create and its DCC launching capabilities.

* [minor] Latest tk-core and tk-desktop2

A minor regression fix with tank validate, and an icon change for tk-desktop2.

* [minor] Updated icon for tk-desktop2.

* Updated to tk-maya engine v0.9.5

Updated to point to the latest release of the tk-maya engine v0.9.5

* Apps, Engines and Core updates

tk-core
- Added support for the Shotgun Unified Login Flow.

tk-houdini
- Compatibility fixes for Houdini 17.

tk-maya
- Fix for window parenting on OSX
- Official support for Maya 2019

tk-nuke
- Updates Nuke version to support 11.3

tk-multi-publish2
- Minor styling change for a button to fix an issue in Houdini 17.

tk-multi-loader2
- Fixes button styling issues evident in Houdini 17.